# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "基于 Gin + Gorm + Wire 的企业级后端服务",
    "title": "Sponge API",
    "termsOfService": "https://github.com/Mojitocean/sponge",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:18888",
  "basePath": "/gateway",
  "paths": {
    "/gateway/chat/message": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "发送聊天消息",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "聊天模块"
        ],
        "summary": "发送消息",
        "parameters": [
          {
            "description": "消息信息",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.SendMessageReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "发送成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.MessageInfoRes"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/gateway/chat/message/list": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取与指定用户的消息列表",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "聊天模块"
        ],
        "summary": "获取消息列表",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "对方用户ID",
            "name": "to_user_id",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 20,
            "description": "每页数量",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/dto.MessageInfoRes"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/gateway/chat/session/list": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取用户的聊天会话列表",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "聊天模块"
        ],
        "summary": "获取会话列表",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 20,
            "description": "每页数量",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/dto.ChatSessionRes"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/gateway/comment/action": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "发布或删除评论",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "评论模块"
        ],
        "summary": "评论操作",
        "parameters": [
          {
            "description": "评论信息",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CommentActionReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.CommentInfoRes"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/gateway/comment/list": {
      "get": {
        "description": "获取视频的评论列表",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "评论模块"
        ],
        "summary": "获取评论列表",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "视频ID",
            "name": "video_id",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 20,
            "description": "每页数量",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/dto.CommentInfoRes"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "dto.ChatSessionRes": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "0"
        },
        "last_message": {
          "type": "string"
        },
        "peer_id": {
          "type": "string",
          "example": "0"
        },
        "peer_user": {
          "description": "对方用户信息",
          "allOf": [
            {
              "$ref": "#/definitions/dto.UserInfoRes"
            }
          ]
        },
        "unread_count": {
          "type": "integer"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
    "dto.CommentActionReq": {
      "type": "object",
      "required": [
        "action",
        "video_id"
      ],
      "properties": {
        "action": {
          "description": "1-发布评论, 2-删除评论",
          "type": "integer",
          "enum": [
            1,
            2
          ]
        },
        "comment_id": {
          "description": "删除时必填",
          "type": "string",
          "example": "0"
        },
        "content": {
          "description": "发布时必填",
          "type": "string"
        },
        "parent_id": {
          "description": "被回复的评论ID（回复时使用）",
          "type": "string",
          "example": "0"
        },
        "root_id": {
          "description": "顶层评论ID（回复时使用）",
          "type": "string",
          "example": "0"
        },
        "video_id": {
          "type": "string",
          "example": "0"
        }
      }
    },
    "dto.CommentInfoRes": {
      "type": "object",
      "properties": {
        "author": {
          "description": "评论作者信息",
          "allOf": [
            {
              "$ref": "#/definitions/dto.UserInfoRes"
            }
          ]
        },
        "content": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "id": {
          "type": "string",
          "example": "0"
        },
        "is_liked": {
          "description": "当前用户是否点赞该评论",
          "type": "boolean"
        },
        "like_count": {
          "type": "integer"
        },
        "parent_id": {
          "type": "string",
          "example": "0"
        },
        "replies": {
          "description": "回复列表（仅顶层评论有）",
          "type": "array",
          "items": {
            "$ref": "#/definitions/dto.CommentInfoRes"
          }
        },
        "reply_count": {
          "description": "回复数量（仅顶层评论有）",
          "type": "integer"
        },
        "root_id": {
          "type": "string",
          "example": "0"
        },
        "user_id": {
          "type": "string",
          "example": "0"
        },
        "video_id": {
          "type": "string",
          "example": "0"
        }
      }
    },
...
    "dto.MessageInfoRes": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "from_id": {
          "type": "string",
          "example": "0"
        },
        "id": {
          "type": "string",
          "example": "0"
        },
        "is_read": {
          "type": "integer"
        },
        "msg_type": {
          "type": "integer"
        },
        "to_id": {
          "type": "string",
          "example": "0"
        }
      }
    },
...
    "dto.SendMessageReq": {
      "type": "object",
      "required": [
        "content",
        "to_user_id"
      ],
      "properties": {
        "content": {
          "type": "string"
        },
        "msg_type": {
          "description": "1-文本, 2-图片",
          "type": "integer",
          "enum": [
            1,
            2
          ]
        },
        "to_user_id": {
          "type": "string",
          "example": "0"
        }
      }
    },
...
    "dto.UserInfoRes": {
      "type": "object",
      "properties": {
        "avatar": {
          "type": "string"
        },
        "backgroundImage": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string",
          "example": "0"
        },
        "nickName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "signature": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        }
      }
    },
...
    "res.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {},
        "msg": {
          "type": "string"
        }
      }
    }
  }
}